+++ test/libsolidity/semanticTests/errors/require_error_function_pointer_parameter.sol
@@ -14,7 +14,5 @@ contract C
     }
 }
 
-// ====
-// compileViaYul: true
 // ----
 // f() -> FAILURE, hex"271b1dfa", hex"c06afe3a8444fc0004668591e8306bfb9968e79ef37cdc8e0000000000000000"